requirements.txt
altair==4.1.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.2.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
blis==0.4.1
Brlapi==0.7.0
cachetools==4.2.4
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.3
command-not-found==0.3
configparser==5.2.0
crayons==0.4.0
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.30
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
easyocr==1.5.0
en-core-web-sm==2.3.1
entrypoints==0.3
et-xmlfile==1.1.0
fake-useragent==0.1.11
fasteners==0.14.1
filelock==3.4.0
fonttools==4.34.4
future==0.18.2
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.24
httplib2==0.14.0
idna==3.3
imageio==2.19.3
importlib-metadata==1.5.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==18.0.1
kiwisolver==1.4.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.7.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
murmurhash==1.0.6
mysql-connector-python==8.0.28
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
networkx==2.8.4
nltk==3.4.5
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
openpyxl==3.0.9
openshot-qt==2.4.3
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plac==1.1.3
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydeck==0.7.1
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==0.9
PyNaCl==1.3.0
pyparsing==3.0.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.0
pytesseract==0.3.9
python-apt==2.0.0+ubuntu0.20.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
regex==2022.3.15
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
sip==4.19.21
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.1
spacy==2.3.1
srsly==1.0.5
streamlit==1.2.0
streamlit-tags==1.2.8
systemd-python==234
terminado==0.12.1
testpath==0.5.0
testresources==2.0.0
thinc==7.4.1
threadpoolctl==3.1.0
tifffile==2022.5.4
toml==0.10.2
toolz==0.11.2
torch==1.12.0
torchvision==0.13.0
tornado==6.1
tqdm==4.62.3
traceback2==1.4.0
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.26.7
usb-creator==0.3.7
useragent==0.1.1
validators==0.18.2
virtualenv==20.0.23
wadllib==1.3.3
wasabi==0.8.2
watchdog==2.1.6
wcwidth==0.2.5
webdriver-manager==3.5.2
webencodings==0.5.1
widgetsnbextension==3.5.2
xkit==0.0.0
zipp==3.6.0